refactor: spending 도메인을 deposit 도메인으로 통합 #25
