+++ src/main/java/cheongsan/domain/deposit/dto/BudgetLimitDTO.java
@@ -1,4 +1,4 @@
-package cheongsan.domain.spending.dto;
+package cheongsan.domain.deposit.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
