+++ src/main/java/cheongsan/domain/deposit/service/BudgetService.java
@@ -1,0 +1,8 @@
+package cheongsan.domain.deposit.service;
+
+import cheongsan.domain.deposit.dto.BudgetLimitDTO;
+
+public interface BudgetService {
+
+    BudgetLimitDTO getBudgetLimits(Long userId);
+}
