+++ src/main/java/cheongsan/domain/deposit/service/DepositService.java
@@ -3,6 +3,7 @@ package cheongsan.domain.deposit.service;
 import cheongsan.domain.deposit.dto.DailyTransactionDTO;
 import cheongsan.domain.deposit.dto.MonthlyTransactionDTO;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -10,4 +11,8 @@ public interface DepositService {
     List<MonthlyTransactionDTO> getMonthlyTransactions(int year, int month);
 
     List<DailyTransactionDTO> getDailyTransactions(LocalDate date);
+
+    BigDecimal calculateRegularMonthlyTransfer(Long userId, int year, int month);
+
+    BigDecimal calculateMonthlyFixedWithdraw(Long userId, int year, int month);
 }
