--- src/main/java/cheongsan/domain/spending/controller/SpendingController.java
@@ -1,4 +1,0 @@
-package cheongsan.domain.spending.controller;
-
-public class SpendingController {
-}
