--- src/main/java/cheongsan/domain/spending/dto/SpendingDTO.java
@@ -1,4 +1,0 @@
-package cheongsan.domain.spending.dto;
-
-public class SpendingDTO {
-}
