--- src/main/java/cheongsan/domain/spending/entity/SpendingEntity.java
@@ -1,4 +1,0 @@
-package cheongsan.domain.spending.entity;
-
-public class SpendingEntity {
-}
