--- src/main/java/cheongsan/domain/spending/mapper/TransactionMapper.java
@@ -1,23 +1,0 @@
-package cheongsan.domain.spending.mapper;
-
-import cheongsan.domain.spending.dto.TransactionDTO;
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Param;
-
-import java.util.List;
-
-@Mapper
-public interface TransactionMapper {
-
-    List<TransactionDTO> findTransferTransactionsByMonth(
-            @Param("userId") Long userId,
-            @Param("year") int year,
-            @Param("month") int month
-    );
-
-    List<TransactionDTO> findWithdrawTransactionsByMonth(
-            @Param("userId") Long userId,
-            @Param("year") int year,
-            @Param("month") int month
-    );
-}
