--- src/main/java/cheongsan/domain/spending/service/BudgetService.java
@@ -1,8 +1,0 @@
-package cheongsan.domain.spending.service;
-
-import cheongsan.domain.spending.dto.BudgetLimitDTO;
-
-public interface BudgetService {
-
-    BudgetLimitDTO getBudgetLimits(Long userId);
-}
