--- src/main/java/cheongsan/domain/spending/service/SpendingService.java
@@ -1,4 +1,0 @@
-package cheongsan.domain.spending.service;
-
-public class SpendingService {
-}
