--- src/main/java/cheongsan/domain/spending/service/TransactionService.java
@@ -1,10 +1,0 @@
-package cheongsan.domain.spending.service;
-
-import java.math.BigDecimal;
-
-public interface TransactionService {
-
-    BigDecimal calculateRegularMonthlyTransfer(Long userId, int year, int month);
-
-    BigDecimal calculateMonthlyFixedWithdraw(Long userId, int year, int month);
-}
