+++ src/main/resources/mapper/DepositMapper.xml
@@ -33,4 +33,41 @@
           AND DATE(t.transaction_time) = #{date}
         ORDER BY t.transaction_time DESC
     </select>
+
+    <select id="findTransferTransactionsByMonth" resultType="cheongsan.domain.deposit.entity.Transaction">
+        SELECT id,
+               deposit_account_id,
+               user_id,
+               transaction_time,
+               after_balance,
+               amount,
+               type,
+               res_account_desc1,
+               res_account_desc2,
+               res_account_desc3
+        FROM transactions
+        WHERE user_id = #{userId}
+          AND type = 'TRANSFER'
+          AND YEAR(transaction_time) = #{year}
+          AND MONTH(transaction_time) = #{month}
+    </select>
+
+    <select id="findWithdrawTransactionsByMonth" resultType="cheongsan.domain.deposit.entity.Transaction">
+        SELECT id,
+               deposit_account_id,
+               user_id,
+               transaction_time,
+               after_balance,
+               amount,
+               type,
+               res_account_desc1,
+               res_account_desc2,
+               res_account_desc3
+        FROM transactions
+        WHERE user_id = #{userId}
+          AND type = 'WITHDRAW'
+          AND YEAR(transaction_time) = #{year}
+          AND MONTH(transaction_time) = #{month}
+    </select>
+
 </mapper>
